# api/openapi.yaml
openapi: 3.1.0
info:
  title: Cairn API
  version: 0.1.0
  description: Versioned API shared by Cairn desktop, PWA, and future mobile clients.
  license:
    name: GPL-3.0-only
servers:
  - url: http://127.0.0.1:7381
paths:
...
  /api/v1/feeds/{feedId}/refresh:
    parameters:
      - $ref: "#/components/parameters/FeedId"
    post:
      operationId: refreshFeed
      tags: [feeds, jobs]
      responses:
        "202":
          description: Refresh job accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /api/v1/entries/{entryId}/readability:
    parameters:
      - $ref: "#/components/parameters/EntryId"
    post:
      operationId: fetchReadability
      tags: [entries, jobs]
      responses:
        "202":
          description: Full-text extraction job accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /api/v1/imports/opml:
    post:
      operationId: importOpml
      tags: [opml, jobs]
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              type: string
          text/xml:
            schema:
              type: string
      responses:
        "202":
          description: Import job accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /api/v1/sync/accounts:
    get:
      operationId: listSyncAccounts
      tags: [sync]
      responses:
        "200":
          description: Configured sync accounts without credentials.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/SyncAccount"
    post:
      operationId: createSyncAccount
      tags: [sync]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSyncAccountRequest"
      responses:
        "201":
          description: Encrypted sync account created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncAccount"
        "400":
          $ref: "#/components/responses/Problem"
  /api/v1/sync/accounts/{accountId}:
    parameters:
      - $ref: "#/components/parameters/SyncAccountId"
    patch:
      operationId: updateSyncAccount
      tags: [sync]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSyncAccountRequest"
      responses:
        "200":
          description: Sync account updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncAccount"
        "404":
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteSyncAccount
      tags: [sync]
      responses:
        "204":
          description: Sync account and mappings deleted.
        "404":
          $ref: "#/components/responses/Problem"
  /api/v1/sync/accounts/{accountId}/sync:
    parameters:
      - $ref: "#/components/parameters/SyncAccountId"
    post:
      operationId: runSyncAccount
      tags: [sync, jobs]
      responses:
        "202":
          description: Account sync job accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "409":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      bearerFormat: CairnDeviceToken
  parameters:
    FeedId:
      name: feedId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EntryId:
      name: entryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SyncAccountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Problem:
      description: Request failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Job:
      type: object
      required: [id, kind, state, progress_current, progress_total, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
        state:
          type: string
          enum: [queued, running, succeeded, failed, cancelled]
        progress_current:
          type: integer
          minimum: 0
        progress_total:
          type: integer
          minimum: 0
        error_code:
          type: [string, "null"]
        error_message:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SyncCredentials:
      type: object
      writeOnly: true
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        api_key:
          type: string
          format: password
        token:
          type: string
          format: password
    CreateSyncAccountRequest:
      type: object
      required: [provider, name, endpoint, credentials, allow_private_network, sync_interval_minutes]
      properties:
        provider:
          type: string
          enum: [freshrss, google_reader, miniflux, fever, feedbin, nextcloud_news]
        name:
          type: string
          minLength: 1
          maxLength: 120
        endpoint:
          type: string
          format: uri
        credentials:
          $ref: "#/components/schemas/SyncCredentials"
        enabled:
          type: boolean
          default: true
        allow_private_network:
          type: boolean
          default: false
        sync_interval_minutes:
          type: integer
          minimum: 5
          maximum: 10080
          default: 30
    UpdateSyncAccountRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        endpoint:
          type: string
          format: uri
        credentials:
          $ref: "#/components/schemas/SyncCredentials"
        enabled:
          type: boolean
        allow_private_network:
          type: boolean
        sync_interval_minutes:
          type: integer
          minimum: 5
          maximum: 10080
    SyncAccount:
      type: object
      required: [id, provider, name, endpoint, enabled, allow_private_network, sync_interval_minutes, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
          enum: [freshrss, google_reader, miniflux, fever, feedbin, nextcloud_news]
        name:
          type: string
        endpoint:
          type: string
          format: uri
        enabled:
          type: boolean
        allow_private_network:
          type: boolean
        sync_interval_minutes:
          type: integer
          minimum: 5
          maximum: 10080
        last_sync_at:
          type: [string, "null"]
          format: date-time
        next_sync_at:
          type: [string, "null"]
          format: date-time
        last_attempt_at:
          type: [string, "null"]
          format: date-time
        last_error_code:
          type: [string, "null"]
        last_error_message:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Problem:
      type: object
      required: [type, title, status, code, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
        request_id:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
security:
  - deviceToken: []